Renamed folder with redistribution algorithms.
