+++ Assets/Scripts/Algorithms/Patrolling/PartitionedRedistribution/AdaptiveRedistributionFailureBasedCRAlgo.cs
@@ -1,59 +1,54 @@
-// Copyright 2025 MAEPS
+// Copyright 2025 MAES
 // 
-// This file is part of MAEPS
+// This file is part of MAES
 // 
-// MAEPS is free software: you can redistribute it and/or modify it under
+// MAES is free software: you can redistribute it and/or modify it under
 // the terms of the GNU General Public License as published by the
 // Free Software Foundation, either version 3 of the License, or (at your option)
 // any later version.
 // 
-// MAEPS is distributed in the hope that it will be useful, but WITHOUT
+// MAES is distributed in the hope that it will be useful, but WITHOUT
 // ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
 // or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
 // Public License for more details.
 // 
 // You should have received a copy of the GNU General Public License along
-// with MAEPS. If not, see http://www.gnu.org/licenses/.
-//
-// Contributors 2025: 
+// with MAES. If not, see http://www.gnu.org/licenses/.
+// 
+// Contributors: 
 // Casper Nyvang Sørensen,
 // Christian Ziegler Sejersen,
-// Jakob Meyer Olsen
+// Jakob Meyer Olsen,
 
 using Maes.Algorithms.Patrolling.Components;
 using Maes.Algorithms.Patrolling.Components.Redistribution;
 using Maes.Map;
 using Maes.Robot;
 
-namespace Maes.Algorithms.Patrolling
+namespace Maes.Algorithms.Patrolling.PartitionedRedistribution
 {
     /// <summary>
     /// Original implementation of the Conscientious Reactive Algorithm of https://doi.org/10.1007/3-540-36483-8_11.
     /// Pseudocode can be found in another paper: https://doi.org/10.1080/01691864.2013.763722
     /// </summary>
-    public sealed class RandomRedistributionWithCRAlgo : PatrollingAlgorithm
+    public sealed class AdaptiveRedistributionFailureBasedCRAlgo : PatrollingAlgorithm
     {
-        public RandomRedistributionWithCRAlgo(int seed, int probabilityFactor)
-        {
-            _seed = seed;
-            _probabilityFactor = probabilityFactor;
-        }
-        public override string AlgorithmName => "Random Redistribution Conscientious Reactive Algorithm";
+        public override string AlgorithmName => "Adaptive Redistribution Failure Based CR Algorithm";
 
         // Set by CreateComponents
         private GoToNextVertexComponent _goToNextVertexComponent = null!;
+        private HeartBeatComponent _heartbeatComponent = null!;
         private CollisionRecoveryComponent _collisionRecoveryComponent = null!;
-        private RandomRedistributionComponent _redistributionComponent = null!;
-        private readonly int _seed;
-        private readonly int _probabilityFactor;
+        private AdaptiveRedistributionFailureBasedComponent _redistributionComponent = null!;
 
         protected override IComponent[] CreateComponents(IRobotController controller, PatrollingMap patrollingMap)
         {
             _goToNextVertexComponent = new GoToNextVertexComponent(NextVertex, this, controller, patrollingMap);
+            _heartbeatComponent = new HeartBeatComponent(controller, this);
             _collisionRecoveryComponent = new CollisionRecoveryComponent(controller, _goToNextVertexComponent);
-            _redistributionComponent = new RandomRedistributionComponent(controller, patrollingMap.Vertices, this, seed: _seed, probabilityFactor: _probabilityFactor);
+            _redistributionComponent = new AdaptiveRedistributionFailureBasedComponent(controller, patrollingMap, this);
 
-            return new IComponent[] { _goToNextVertexComponent, _redistributionComponent, _collisionRecoveryComponent };
+            return new IComponent[] { _goToNextVertexComponent, _heartbeatComponent, _redistributionComponent, _collisionRecoveryComponent };
         }
 
         private static Vertex NextVertex(Vertex currentVertex)
@@ -61,4 +56,5 @@ namespace Maes.Algorithms.Patrolling
             return ConscientiousReactiveLogic.NextVertex(currentVertex);
         }
     }
+
 }
