+++ Assets/Scripts/Algorithms/Patrolling/PartitionedRedistribution/RandomRedistributionWithCRAlgo.cs
@@ -1,52 +1,59 @@
-// Copyright 2025 MAES
+// Copyright 2025 MAEPS
 // 
-// This file is part of MAES
+// This file is part of MAEPS
 // 
-// MAES is free software: you can redistribute it and/or modify it under
+// MAEPS is free software: you can redistribute it and/or modify it under
 // the terms of the GNU General Public License as published by the
 // Free Software Foundation, either version 3 of the License, or (at your option)
 // any later version.
 // 
-// MAES is distributed in the hope that it will be useful, but WITHOUT
+// MAEPS is distributed in the hope that it will be useful, but WITHOUT
 // ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
 // or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
 // Public License for more details.
 // 
 // You should have received a copy of the GNU General Public License along
-// with MAES. If not, see http://www.gnu.org/licenses/.
-// 
-// Contributors: 
+// with MAEPS. If not, see http://www.gnu.org/licenses/.
+//
+// Contributors 2025: 
 // Casper Nyvang Sørensen,
 // Christian Ziegler Sejersen,
-// Jakob Meyer Olsen,
+// Jakob Meyer Olsen
 
 using Maes.Algorithms.Patrolling.Components;
 using Maes.Algorithms.Patrolling.Components.Redistribution;
 using Maes.Map;
 using Maes.Robot;
 
-namespace Maes.Algorithms.Patrolling.PartitionedAlgorithms
+namespace Maes.Algorithms.Patrolling.PartitionedRedistribution
 {
     /// <summary>
     /// Original implementation of the Conscientious Reactive Algorithm of https://doi.org/10.1007/3-540-36483-8_11.
     /// Pseudocode can be found in another paper: https://doi.org/10.1080/01691864.2013.763722
     /// </summary>
-    public sealed class AdaptiveRedistributionSuccessBasedCRAlgo : PatrollingAlgorithm
+    public sealed class RandomRedistributionWithCRAlgo : PatrollingAlgorithm
     {
-        public override string AlgorithmName => "Adaptive Redistribution Success Based CR Algorithm";
+        public RandomRedistributionWithCRAlgo(int seed, int probabilityFactor)
+        {
+            _seed = seed;
+            _probabilityFactor = probabilityFactor;
+        }
+        public override string AlgorithmName => "Random Redistribution Conscientious Reactive Algorithm";
 
         // Set by CreateComponents
         private GoToNextVertexComponent _goToNextVertexComponent = null!;
         private CollisionRecoveryComponent _collisionRecoveryComponent = null!;
-        private AdaptiveRedistributionSuccessBasedComponent _redistributionComponent = null!;
+        private RandomRedistributionComponent _redistributionComponent = null!;
+        private readonly int _seed;
+        private readonly int _probabilityFactor;
 
         protected override IComponent[] CreateComponents(IRobotController controller, PatrollingMap patrollingMap)
         {
             _goToNextVertexComponent = new GoToNextVertexComponent(NextVertex, this, controller, patrollingMap);
             _collisionRecoveryComponent = new CollisionRecoveryComponent(controller, _goToNextVertexComponent);
-            _redistributionComponent = new AdaptiveRedistributionSuccessBasedComponent(controller, patrollingMap, this);
+            _redistributionComponent = new RandomRedistributionComponent(controller, patrollingMap.Vertices, this, seed: _seed, probabilityFactor: _probabilityFactor);
 
-            return new IComponent[] { _goToNextVertexComponent, _collisionRecoveryComponent, _redistributionComponent };
+            return new IComponent[] { _goToNextVertexComponent, _redistributionComponent, _collisionRecoveryComponent };
         }
 
         private static Vertex NextVertex(Vertex currentVertex)
@@ -54,5 +61,4 @@ namespace Maes.Algorithms.Patrolling.PartitionedAlgorithms
             return ConscientiousReactiveLogic.NextVertex(currentVertex);
         }
     }
-
 }
