+++ Assets/Scripts/Experiments/Patrolling/PartitionedConscientiousReactiveExperiment.cs
@@ -26,6 +26,7 @@
 using System.Collections.Generic;
 
 using Maes.Algorithms.Patrolling.PartitionedAlgorithms;
+using Maes.Algorithms.Patrolling.PartitionedRedistribution;
 using Maes.Map.Generators;
 using Maes.Robot;
 using Maes.Simulation.Patrolling;
